Increase coverage for core module
